api/requirements.txt
# Core Dependencies
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
llama-cpp-python
openai

# Audio Transcription
faster-whisper
